+++ nishiki_koi_shop/src/components/Header.js
@@ -63,6 +63,7 @@ const Header = () => {
 
     const handleSignOut = () => {
         sessionStorage.removeItem('user');
+        handleCloseNavBar();
     }
 
     useEffect(() => {
@@ -146,12 +147,7 @@ const Header = () => {
                                 {list.map((item, index) => (
                                     <li key={1 * index} className={isActive ? 'show' : 'hidden'}>
                                         {item === 'Đăng xuất'
-
-                                            ? (<Link to={'/'} onClick={() => handleSignOut()}>{item} </Link>)
-                                            : (<Link to={`/${linkMenu[3][index]}`}>{item}</Link>)
-
-                                            ? (<Link to={'/'} onClick={() => handleSignOut()}
-                                                     onClick={handleCloseNavBar}>{item} < /Link>)
+                                            ? (<Link to={'/'} onClick={() => handleSignOut()}>{item} < /Link>)
                                             : (<Link to={`/${linkMenu[1][index]}`}
                                                      onClick={handleCloseNavBar}>{item}</Link>)
 
@@ -164,23 +160,15 @@ const Header = () => {
                         <li
                             onMouseEnter={() => HandleMouseEnter(['Đăng nhập', 'Đăng kí'])}
                             onMouseLeave={() => HandleMouseLeave()}
-
-                        >
-                            <Link to={'/account'}>Tài khoản</Link>
+                        ><Link to={'/account'} onClick={handleCloseNavBar}>Tài khoản</Link>
                             <ul className={'sub-nav'}>
                                 {list.map((item, index) => (
                                     <li key={5 * index} className={isActive ? 'show' : 'hidden'}>
                                         {item === 'Đăng kí' ? (
                                             <Link to={'/register'}>{item}</Link> // Điều hướng đến trang đăng ký
                                         ) : (
-                                            <Link to={`/${linkMenu[4][index]}`}>{item}</Link>
+                                            <Link to={`/${linkMenu[2][index]}`}>{item}</Link>
                                         )}
-
-                        ><Link to={'/account'} onClick={handleCloseNavBar}>Tài khoản</Link>
-                            <ul className={'sub-nav'}>
-                                {list.map((item, index) => (
-                                    <li key={2 * index} className={isActive ? 'show' : 'hidden'}>
-                                        <Link to={`/${linkMenu[2    ][index]}`} onClick={handleCloseNavBar}>{item}</Link>
                                     </li>
                                 ))}
                             </ul>
