+++ nishiki_koi_shop/src/pages/Manager/Manager.js
