+++ nishiki_koi_shop/src/pages/Manager/Manager_Trip.js
